clean up some things start g64 support add LZ77 coder/decoder

+++ lib/meatloaf/meat_buffer.h
@@ -198,7 +198,7 @@ namespace Meat
                 else
                 {
                     currBuffEnd = mstream->position();
-                    currBuffStart = currBuffEnd - readCount; // this is where our buffer data starts
+                    currBuffStart = (uint32_t)currBuffEnd - readCount; // this is where our buffer data starts
 
                     // Debug_printv("--mfilebuf underflow, read bytes=%d--", readCount);
                     // beg, curr, end <=> eback, gptr, egptr
@@ -255,7 +255,7 @@ namespace Meat
 
             char *end = this->pptr();
 
-            Debug_printv("before write call, ch=%d [%c], buffer contains:%d", ch, (char)ch, this->pptr() - this->pbase());
+            Debug_printv("before write call, ch=%d [%c], buffer contains:%lu", ch, (char)ch, this->pptr() - this->pbase());
 
             if (ch != EOF)
             {
@@ -263,7 +263,7 @@ namespace Meat
                 this->pbump(1);
             }
 
-            Debug_printv("%d bytes in buffer will be written", end - this->pbase());
+            Debug_printv("%lu bytes in buffer will be written", end - this->pbase());
 
             const uint8_t *pBase = (uint8_t *)this->pbase();
 
@@ -306,7 +306,7 @@ namespace Meat
             // gptr - current gbuffer position
             // egptr - gbuffer end
 
-            Debug_printv("meat buffer seekpos called, newPos=%d buffer=[%d,%d]", (size_t)__pos, (size_t)currBuffStart, (size_t)currBuffEnd);
+            Debug_printv("meat buffer seekpos called, newPos=%d buffer=[%d,%d]", (int)__pos, (int)currBuffStart, (int)currBuffEnd);
 
             if (__pos >= currBuffStart && __pos < currBuffEnd)
             {
